// common/enums/SkatteverketEnum.ts
export enum SkatteverketStatusEnum {
  STARTED = 'SKATTEVERKET_STATUS_STARTED',
  BANK_ID_REQUIRED = 'SKATTEVERKET_STATUS_BANK_ID_REQUIRED',
  CANCELLED = 'SKATTEVERKET_STATUS_CANCELLED',
  PENDING = 'SKATTEVERKET_STATUS_PENDING',
  COMPLETED = 'SKATTEVERKET_STATUS_COMPLETED',
  TOKEN = 'skatteverketToken',
  STATUS_OK = 'OK',
  EMPTY = 'SKATTEVERKET_STATUS_EMPTY',
  STATUS = 'SKATTEVERKET_STATUS',
  APARTMENTS = 'LAGENHETER_FINNS',
  NO_MATCH = 'EJ_MATCH_FINNS_ANDRA',
  NO_ADDRESSES_FOUND = 'NO_ADDRESSES_FOUND',
}

export enum SkatteverketErrorEnum {
  BANKID_TIMEOUT = 'SKATTEVERKET_BANKID_TIMEOUT',
  TIMEOUT = 'SKATTEVERKET_TIMEOUT',
  USER_TIMEOUT = 'SKATTEVERKET_USER_TIMEOUT',
  COOKIE_NOT_FOUND = 'SKATTEVERKET_COOKIE_NOT_FOUND',
}

export enum SkatteverketActionsEnum {
  SET_SEARCH_ADDRESS_RESPONSE = 'SET_SEARCH_ADDRESS_RESPONSE',
  SET_CURRENT_ADDRESS = 'SET_CURRENT_ADDRESS',
  SET_PERSONS_AT_CURRENT_ADDRESS = 'SET_PERSONS_AT_CURRENT_ADDRESS',
  SET_newAddress = 'SET_newAddress',
  SET_ADDRESS_VALIDATION = 'SET_ADDRESS_VALIDATION',
  SET_AVAILABLE_APARTMENTNUMBERS = 'SET_AVAILABLE_APARTMENTNUMBERS',
  SET_IS_BANKID_LOADING = 'SET_IS_BANKID_LOADING',
  SET_IS_SKATTEVERKET_LOADING = 'SET_IS_SKATTEVERKET_LOADING',
  SET_SKATTEVERKET_LOGIN_STATUS = 'SET_SKATTEVERKET_LOGIN_STATUS',
}